BeforeProfileMigration

+++ GraduationProject/Abstract/IAllOrders.cs
@@ -5,6 +5,6 @@ namespace GraduationProject.Abstract
     public interface IAllOrders
     {
         IQueryable<openOrder> Orders { get;}
-        openOrder GetObjectOrder(int orderId);
+        openOrder GetObjectOrder(long orderId);
     }
 }
